splitted sourceBuffer's logic;start split mediaSource #6091
